backend/requirements.txt
PyYAML
websocket-server
